Charts for 7 day forecast are working

+++ ApiLibrary/WeatherTempModel.cs
@@ -8,9 +8,9 @@ namespace ApiLibrary
 {
     public class WeatherTempModel
     {
-        public float Day { get; set; }
-        public float Night { get; set; }
-        public float Eve { get; set; }
-        public float Morn { get; set; }
+        public double Day { get; set; }
+        public double Night { get; set; }
+        public double Eve { get; set; }
+        public double Morn { get; set; }
     }
 }
